+++ lib/sdk_flutter_platform_interface.dart
@@ -3,27 +3,37 @@ import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'sdk_flutter_method_channel.dart';
 
 abstract class SdkFlutterPlatform extends PlatformInterface {
-  /// Constructs a SdkFlutterPlatform.
   SdkFlutterPlatform() : super(token: _token);
 
   static final Object _token = Object();
 
   static SdkFlutterPlatform _instance = MethodChannelSdkFlutter();
 
-  /// The default instance of [SdkFlutterPlatform] to use.
-  ///
-  /// Defaults to [MethodChannelSdkFlutter].
   static SdkFlutterPlatform get instance => _instance;
 
-  /// Platform-specific implementations should set this with their own
-  /// platform-specific class that extends [SdkFlutterPlatform] when
-  /// they register themselves.
   static set instance(SdkFlutterPlatform instance) {
     PlatformInterface.verifyToken(instance, _token);
     _instance = instance;
   }
 
-  Future<String?> getPlatformVersion() {
-    throw UnimplementedError('platformVersion() has not been implemented.');
+  /// Start verification with an existing session token.
+  Future<Map<String, dynamic>> startVerification(
+    String token,
+    Map<String, dynamic>? config,
+  ) {
+    throw UnimplementedError('startVerification() has not been implemented.');
+  }
+
+  /// Start verification by creating a session with a workflow ID.
+  Future<Map<String, dynamic>> startVerificationWithWorkflow(
+    String workflowId,
+    String? vendorData,
+    String? metadata,
+    Map<String, dynamic>? contactDetails,
+    Map<String, dynamic>? expectedDetails,
+    Map<String, dynamic>? config,
+  ) {
+    throw UnimplementedError(
+        'startVerificationWithWorkflow() has not been implemented.');
   }
 }